feat(frontend): rearrange options more logically
this will also keep layout and display mostly the same length reducing need of scrolling. now layout is more used for arranging how things will appear, display is more about turning on / off different fields or field specific features, and since font sliders can apply to multiple fields it seems to be okay in the layout section. organized in a way that hopefully the most important or useful settings are at the top.