feat(frontend): match line-heights
hindi line-height is now the same as punjabi. urdu defaults to 2.52 which is quite a bit. testing at 2.1 seemed to have an extremely minimal chance to overlap. time will tell if we need to widen this. set to something similar to punjabi/hindi line-heights. word spacing added for readability, word gaps about the same or a little bigger than one dash of the urdu full-stop character.